[BE] refactor: make Prisma option refactor
option 객체를 만드는 메소드 리팩토링,
중복 코드 제거
